colorize: extract color primitives into a leaf package
Option 1 of #933: ColorMode, its parsing and validation, the
color-enabled switch, and piped-output detection move to a dedicated
leaf package owned by neither the checker nor the formatters. The
rules package uses it for severity rendering, formatters take it
directly instead of reaching back into checker, and the checker
package keeps aliases so its public surface is unchanged.

The larger half of #933 (moving Change rendering itself into
formatters) remains a separate redesign, as the issue suggests.
