+++ colorize/piped_output.go
@@ -1,4 +1,4 @@
-package rules
+package colorize
 
 import "os"
